requirements.txt
kafka-python
confluent-kafka
apache-airflow-providers-docker
asttokens
comm
contourpy
cycler
debugpy
decorator
exceptiongroup
executing
fonttools
ipykernel
ipython
jedi
joblib
jupyter_client
jupyter_core
kiwisolver
matplotlib
matplotlib-inline
nest-asyncio
numpy
nvidia-nccl-cu12
packaging
pandas
parso
pexpect
pillow
platformdirs
prompt_toolkit
psutil
ptyprocess
pure_eval
Pygments
pyparsing
python-dateutil
pytz
pyzmq
scikit-learn
scipy
seaborn
six
stack-data
threadpoolctl
tornado
traitlets
typing_extensions
tzdata
wcwidth
xgboost